RF: move assert_arr_dict_equal to testing
Move testing function from parrect tests to general testing utilities.
